refactor: polish code
